Split tests in signup (> 1k lines)
