+++ itou/www/signup/tests/test_allauth.py
@@ -1,0 +1,14 @@
+from django.test import TestCase
+from django.urls import reverse
+
+
+class AllauthSignupTest(TestCase):
+    def test_allauth_signup_url_override(self):
+        """Ensure that the default allauth signup URL is overridden."""
+        ALLAUTH_SIGNUP_URL = reverse("account_signup")
+        self.assertEqual(ALLAUTH_SIGNUP_URL, "/accounts/signup/")
+        response = self.client.get(ALLAUTH_SIGNUP_URL)
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "signup/signup.html")
+        response = self.client.post(ALLAUTH_SIGNUP_URL, data={"foo": "bar"})
+        self.assertEqual(response.status_code, 405)
