+++ itou/www/signup/tests/test_prescriber.py
@@ -3,286 +3,24 @@ from unittest import mock
 
 import httpx
 import respx
-from allauth.account.forms import default_token_generator
 from allauth.account.models import EmailConfirmationHMAC
-from allauth.account.utils import user_pk_to_url_str
 from django.conf import settings
 from django.core import mail
 from django.test import TestCase
 from django.urls import reverse
-from django.utils.html import escape
-from django.utils.http import urlencode
 
-from itou.cities.factories import create_test_cities
-from itou.cities.models import City
 from itou.prescribers.factories import (
     PrescriberOrganizationFactory,
     PrescriberOrganizationWithMembershipFactory,
     PrescriberPoleEmploiFactory,
 )
 from itou.prescribers.models import PrescriberOrganization
-from itou.siaes.factories import SiaeFactory
-from itou.siaes.models import Siae
-from itou.users.factories import DEFAULT_PASSWORD, JobSeekerFactory
+from itou.users.factories import DEFAULT_PASSWORD
 from itou.users.models import User
 from itou.utils.mocks.api_entreprise import ETABLISSEMENT_API_RESULT_MOCK
 from itou.utils.mocks.geocoding import BAN_GEOCODING_API_RESULT_MOCK
 from itou.utils.password_validation import CnilCompositionPasswordValidator
-from itou.www.signup.forms import JobSeekerSituationForm, PrescriberChooseKindForm
-
-
-class SignupTest(TestCase):
-    def test_allauth_signup_url_override(self):
-        """Ensure that the default allauth signup URL is overridden."""
-        ALLAUTH_SIGNUP_URL = reverse("account_signup")
-        self.assertEqual(ALLAUTH_SIGNUP_URL, "/accounts/signup/")
-        response = self.client.get(ALLAUTH_SIGNUP_URL)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "signup/signup.html")
-        response = self.client.post(ALLAUTH_SIGNUP_URL, data={"foo": "bar"})
-        self.assertEqual(response.status_code, 405)
-
-
-class SiaeSignupTest(TestCase):
-    def test_join_an_siae_without_members(self):
-        """
-        A user joins an SIAE without members.
-
-        The full "email confirmation process" is tested here.
-        Further Siae's signup tests doesn't have to fully test it again.
-        """
-
-        user_first_name = "Jacques"
-        user_email = "jacques.doe@siae.com"
-        user_secondary_email = "jacques.doe@hotmail.com"
-
-        siae = SiaeFactory(kind=Siae.KIND_ETTI)
-        self.assertEqual(0, siae.members.count())
-
-        token = siae.get_token()
-        with mock.patch("itou.utils.tokens.SiaeSignupTokenGenerator.make_token", return_value=token):
-
-            url = reverse("signup:siae_select")
-            response = self.client.get(url)
-            self.assertEqual(response.status_code, 200)
-
-            # Find an SIAE by SIREN.
-            response = self.client.get(url, {"siren": siae.siret[:9]})
-            self.assertEqual(response.status_code, 200)
-
-            # Choose an SIAE between results.
-            post_data = {"siaes": siae.pk}
-            # Pass `siren` in request.GET
-            response = self.client.post(f"{url}?siren={siae.siret[:9]}", data=post_data)
-            self.assertEqual(response.status_code, 302)
-            self.assertRedirects(response, "/")
-
-            self.assertEqual(len(mail.outbox), 1)
-            email = mail.outbox[0]
-            self.assertIn("Un nouvel utilisateur souhaite rejoindre votre structure", email.subject)
-
-            magic_link = siae.signup_magic_link
-            response = self.client.get(magic_link)
-            self.assertEqual(response.status_code, 200)
-
-            # No error when opening magic link a second time.
-            response = self.client.get(magic_link)
-            self.assertEqual(response.status_code, 200)
-
-            # Create user.
-            url = siae.signup_magic_link
-            post_data = {
-                # Hidden fields
-                "encoded_siae_id": siae.get_encoded_siae_id(),
-                "token": siae.get_token(),
-                # Readonly fields
-                "siret": siae.siret,
-                "kind": siae.kind,
-                "siae_name": siae.display_name,
-                # Regular fields
-                "first_name": user_first_name,
-                "last_name": "Doe",
-                "email": user_secondary_email,
-                "password1": DEFAULT_PASSWORD,
-                "password2": DEFAULT_PASSWORD,
-            }
-            response = self.client.post(url, data=post_data)
-            self.assertEqual(response.status_code, 302)
-            self.assertRedirects(response, reverse("account_email_verification_sent"))
-
-            self.assertFalse(User.objects.filter(email=user_email).exists())
-            user = User.objects.get(email=user_secondary_email)
-
-            # Check `User` state.
-            self.assertFalse(user.is_job_seeker)
-            self.assertFalse(user.is_prescriber)
-            self.assertTrue(user.is_siae_staff)
-            self.assertTrue(user.is_active)
-            self.assertTrue(siae.has_admin(user))
-            self.assertEqual(1, siae.members.count())
-            # `username` should be a valid UUID, see `User.generate_unique_username()`.
-            self.assertEqual(user.username, uuid.UUID(user.username, version=4).hex)
-            self.assertEqual(user.first_name, user_first_name)
-            self.assertEqual(user.last_name, post_data["last_name"])
-            self.assertEqual(user.email, user_secondary_email)
-            # Check `EmailAddress` state.
-            self.assertEqual(user.emailaddress_set.count(), 1)
-            user_email = user.emailaddress_set.first()
-            self.assertFalse(user_email.verified)
-
-            # Check sent email.
-            self.assertEqual(len(mail.outbox), 2)
-            subjects = [email.subject for email in mail.outbox]
-            self.assertIn("[Action requise] Un nouvel utilisateur souhaite rejoindre votre structure !", subjects)
-            self.assertIn("Confirmez votre adresse e-mail", subjects)
-
-            # Magic link is no longer valid because siae.members.count() has changed.
-            response = self.client.get(magic_link, follow=True)
-            redirect_url, status_code = response.redirect_chain[-1]
-            self.assertEqual(status_code, 302)
-            next_url = reverse("signup:siae_select")
-            self.assertEqual(redirect_url, next_url)
-            self.assertEqual(response.status_code, 200)
-            expected_message = (
-                "Ce lien d'inscription est invalide ou a expiré. " "Veuillez procéder à une nouvelle inscription."
-            )
-            self.assertContains(response, escape(expected_message))
-
-            # User cannot log in until confirmation.
-            post_data = {"login": user.email, "password": DEFAULT_PASSWORD}
-            url = reverse("account_login")
-            response = self.client.post(url, data=post_data)
-            self.assertEqual(response.status_code, 302)
-            self.assertEqual(response.url, reverse("account_email_verification_sent"))
-
-            # Confirm email + auto login.
-            confirmation_token = EmailConfirmationHMAC(user_email).key
-            confirm_email_url = reverse("account_confirm_email", kwargs={"key": confirmation_token})
-            response = self.client.post(confirm_email_url)
-            self.assertEqual(response.status_code, 302)
-            self.assertEqual(response.url, reverse("welcoming_tour:index"))
-            user_email = user.emailaddress_set.first()
-            self.assertTrue(user_email.verified)
-
-
-class JobSeekerSignupTest(TestCase):
-    def setUp(self):
-        create_test_cities(["67"], num_per_department=1)
-
-    def test_job_seeker_signup_situation(self):
-        """
-        Test the redirects according to the chosen situations
-        """
-
-        # Check if the form page is displayed correctly.
-        url = reverse("signup:job_seeker_situation")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-
-        # Check if none of the boxes are checked 'some data' needed to raise
-        # form error.
-        post_data = {"some": "data"}
-        response = self.client.post(url, post_data)
-        self.assertEqual(response.status_code, 200)
-        self.assertFormError(response, "form", "situation", [JobSeekerSituationForm.ERROR_NOTHING_CHECKED])
-
-        # Check if one of eligibility criterion is checked.
-        next_url = reverse("signup:job_seeker")
-        for choice in JobSeekerSituationForm.ELIGIBLE_SITUATION:
-            post_data = {"situation": [choice]}
-            response = self.client.post(url, data=post_data)
-            self.assertEqual(response.status_code, 302)
-            self.assertRedirects(response, next_url)
-
-            post_data["situation"].append("autre")
-            response = self.client.post(url, data=post_data)
-            self.assertEqual(response.status_code, 302)
-            self.assertRedirects(response, next_url)
-
-        # Check if all the eligibility criteria are checked.
-        post_data = {"situation": JobSeekerSituationForm.ELIGIBLE_SITUATION}
-        response = self.client.post(url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-        self.assertRedirects(response, next_url)
-
-        # Check if "Autre" is the only one checked.
-        post_data = {"situation": "autre"}
-        response = self.client.post(url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-        next_url = reverse("signup:job_seeker_situation_not_eligible")
-        self.assertRedirects(response, next_url)
-
-        # Check not eligible destination page.
-        url = reverse("signup:job_seeker_situation_not_eligible")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-
-    def test_job_seeker_signup(self):
-        """Job-seeker signup."""
-
-        url = reverse("signup:job_seeker")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-
-        address_line_1 = "Test adresse"
-        address_line_2 = "Test adresse complémentaire"
-        city = City.objects.first()
-        post_code = city.post_codes[0]
-
-        post_data = {
-            "first_name": "John",
-            "last_name": "Doe",
-            "email": "john.doe@siae.com",
-            "password1": DEFAULT_PASSWORD,
-            "password2": DEFAULT_PASSWORD,
-            "address_line_1": address_line_1,
-            "address_line_2": address_line_2,
-            "post_code": post_code,
-            "city_name": city.name,
-            "city": city.slug,
-        }
-
-        response = self.client.post(url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-        self.assertRedirects(response, reverse("account_email_verification_sent"))
-
-        # Check `User` state.
-        user = User.objects.get(email=post_data["email"])
-        # `username` should be a valid UUID, see `User.generate_unique_username()`.
-        self.assertEqual(user.username, uuid.UUID(user.username, version=4).hex)
-        self.assertTrue(user.is_job_seeker)
-        self.assertFalse(user.is_prescriber)
-        self.assertFalse(user.is_siae_staff)
-
-        # Check `EmailAddress` state.
-        self.assertEqual(user.emailaddress_set.count(), 1)
-        user_email = user.emailaddress_set.first()
-        self.assertFalse(user_email.verified)
-
-        # Check sent email.
-        self.assertEqual(len(mail.outbox), 1)
-        email = mail.outbox[0]
-        self.assertIn("Confirmez votre adresse e-mail", email.subject)
-        self.assertIn("Afin de finaliser votre inscription, cliquez sur le lien suivant", email.body)
-        self.assertEqual(email.from_email, settings.DEFAULT_FROM_EMAIL)
-        self.assertEqual(len(email.to), 1)
-        self.assertEqual(email.to[0], user.email)
-
-        # User cannot log in until confirmation.
-        post_data = {"login": user.email, "password": DEFAULT_PASSWORD}
-        url = reverse("account_login")
-        response = self.client.post(url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse("account_email_verification_sent"))
-
-        # Confirm email + auto login.
-        confirmation_token = EmailConfirmationHMAC(user_email).key
-        confirm_email_url = reverse("account_confirm_email", kwargs={"key": confirmation_token})
-        response = self.client.post(confirm_email_url)
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse("welcoming_tour:index"))
-        user_email = user.emailaddress_set.first()
-        self.assertTrue(user_email.verified)
+from itou.www.signup.forms import PrescriberChooseKindForm
 
 
 class PrescriberSignupTest(TestCase):
@@ -944,85 +682,3 @@ class PrescriberSignupTest(TestCase):
         }
         response = self.client.post(url, data=post_data)
         self.assertContains(response, f"« {prescriber_organization.display_name} » utilise déjà ce SIRET.")
-
-
-class PasswordResetTest(TestCase):
-    def test_password_reset_flow(self):
-
-        user = JobSeekerFactory()
-
-        # Ask for password reset.
-        url = reverse("account_reset_password")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-
-        post_data = {"email": user.email}
-        response = self.client.post(url, data=post_data)
-        args = urlencode({"email": user.email})
-        next_url = reverse("account_reset_password_done")
-        self.assertRedirects(response, f"{next_url}?{args}")
-
-        # Check sent email.
-        self.assertEqual(len(mail.outbox), 1)
-        email = mail.outbox[0]
-        self.assertIn("Réinitialisation de votre mot de passe", email.subject)
-        self.assertIn(
-            "Si vous n'avez pas demandé la réinitialisation de votre mot de passe, vous pouvez ignorer ce message",
-            email.body,
-        )
-        self.assertEqual(email.from_email, settings.DEFAULT_FROM_EMAIL)
-        self.assertEqual(len(email.to), 1)
-        self.assertEqual(email.to[0], user.email)
-
-        # Change forgotten password.
-        uidb36 = user_pk_to_url_str(user)
-        key = default_token_generator.make_token(user)
-        password_change_url = reverse("account_reset_password_from_key", kwargs={"uidb36": uidb36, "key": key})
-        response = self.client.get(password_change_url)
-        password_change_url_with_hidden_key = response.url
-        post_data = {"password1": DEFAULT_PASSWORD, "password2": DEFAULT_PASSWORD}
-        response = self.client.post(password_change_url_with_hidden_key, data=post_data)
-        self.assertRedirects(response, reverse("account_reset_password_from_key_done"))
-
-        # User can log in with his new password.
-        self.assertTrue(self.client.login(username=user.email, password=DEFAULT_PASSWORD))
-        self.client.logout()
-
-    def test_password_reset_with_nonexistent_email(self):
-        """
-        Avoid user enumeration: redirect to the success page even with a nonexistent email.
-        """
-        url = reverse("account_reset_password")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-        post_data = {"email": "nonexistent@email.com"}
-        response = self.client.post(url, data=post_data)
-        args = urlencode({"email": post_data["email"]})
-        next_url = reverse("account_reset_password_done")
-        self.assertEqual(response.url, f"{next_url}?{args}")
-
-
-class PasswordChangeTest(TestCase):
-    def test_password_change_flow(self):
-        """
-        Ensure that the default allauth account_change_password URL is overridden
-        and redirects to the right place.
-        """
-
-        user = JobSeekerFactory()
-        self.assertTrue(self.client.login(username=user.email, password=DEFAULT_PASSWORD))
-
-        # Change password.
-        url = reverse("account_change_password")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-        new_password = "Mlkjhgf!sq2"
-        post_data = {"oldpassword": DEFAULT_PASSWORD, "password1": new_password, "password2": new_password}
-        response = self.client.post(url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-        self.assertRedirects(response, reverse("dashboard:index"))
-
-        # User can log in with his new password.
-        self.client.logout()
-        self.assertTrue(self.client.login(username=user.email, password=new_password))
-        self.client.logout()
